Player Tail Swipe checking now works, force is not applied to anything.
Also made the player a prefab.
