+++ Assets/Scripts/Player/TailSwipeCollision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9737b59e73706d488c6022c4f507a3f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
